stageMover.makeInitialPublications: don't publish 'stage step size' event.
The 'stage step size' event is for when the step size changes.
At this point it's not changing it's just emitting the initial
step size after construction.  This event is subscribed by GUI
elements that display the current step size but they don't need
this either since they default to the step size value when they
are constructed.