MacroStageWindow: Centre() instead of CentreVertical() for wxWidgets<3.1.
Centre the row of buttons in the middle of the macro stage window
instead of centre vertical.  It looks the same since the row of
buttons does not expand and is on a vertical box sizer but avoids the
'wx.SizerVertical.CentreHorizontal' method which was only added
wxWidgets 3.1.0 (wxPython 4.1.0 or later).