Replace wrapping FTGL with our own port wrapping freetype-py (issue #615)
FTGL is our main source of installation issues because there are no
official binaries and is abandoned by upstream.  The two python
packages that wrap it are also abandoned.

This commit replaces our use of FTGL by creating our own Python port
of FTGL, limited to our needs.  FTGL wraps Freetype2 for use with
OpenGL.  Instead of the Freetype2 C library, we use freetype-py.
Because this commit is a port of FTGL, we keep its original copyright.