MacroStageWindow: split into MacroStagePanel to fix background on Windows.
On Windows we should place controls on a panel which is then placed
on the frame.  If there is no panel then the background will be a
dark grey instead of the typical light grey from the native dialog
windows.  See also 73d1b6c28126 for a similar fix on the main window.