macroStageWindow.AxisStepCtrl: enable manual edit of the step size (issue #612)
This commit enables the ability for the user to set the step size to
any value they like.  Besides giving the user more flexibility, this
is more important since f84cc8171800 which removed limits on the step
size.  If the user accidentally increased/decreased step size, this
provides a method to quickly set it back to something sensible.