PositionerHandler: remove cleanupAfterExperiment method and callback (#635)
The `cleanupAfterExperiment` callback is for stages eligible for
experiments.  It is mainly used to revert the stage back to its
position before the experiment started.  However, this is already done
by Experiment.cleanup and most stages do nothing here.

This commit removes PositionerHandler.cleanupAfterExperiment (there's
an empty fallback on DeviceHandler) and the cleanupAfterExperiment
callbacks on the stage devices.