DrawerHandler: set initial drawer setting instead of zero when cokpit starts.
When cockpit starts, changeDrawer is called without arguments and
defaults to the first settings in the list.  However, DrawerHandler
when constructed takes an index of that list to mark what should be
the inital value.

This commit uses that initial value when calling changeDrawer at the
end of cockpit initialization.