cockpit.gui.toggleButton: remove no longer used implementation of ToggleButton.
Cockpit used to implement most buttons with its own implementation of
a ToggleButton which is actually a wx.StaticText with a coloured
background.  We have slowly migrating to wx native controls to achieve
an interface that will look native to the system.

cockpit.gui.toggleButton.ToggleButton is now no longer used in Cockpit
and this commit removes it.