Set light power with fractions of power instead of mW (#641)
Using physical units mW to set laser power can be misleading because
that's the laser power in the device and not the power that reaches
the sample (and that's even assuming that the laser is properly
calibrated).  Python Microscope just changed its interface to use
fractional values (in the [0, 1] interval) instead of mW.

This commit adjusts the MicroscopeLaser device to the new upstream
interface; it adjusts LightPowerHandler which no longer requires min
and max power; and adjusts the touchscreen and main window panels.