WindowsMenu: merge "Move to mouse" into the "Show" menu item (#652)
The "Move to mouse" menu item in the windows menu is for when the position of
cockpit window is outside the display, typically because a display was removed
or added.  For the user point of view, the window is just hidden and the "Show"
menu item should make it appear, whether the window is hidden because it's not
wx.Shown, or because it's outside the display is irrelevant.

This commit removes the "Move to "mouse" menu item, merging its function into
the "Show" menu item.