WindowMenu: replace window submenus with a single menu item (#652)
There used to be multiple actions for each window on the menu bar, hence the
submenu per window.  However, there is now a single action, "Show", so a submenu
for each window is no longer sensible.

This commit replaces the submenu for each window with a single menu item named
after the window title.  This menu item makes the window appear whether it is
hidden, minimised, under other windows, or outside the display.