EnumerateSitesPanel: fix parsing site list in Python 3 (#662)
In Python 2, filter() would return a list but since Python 3, filter()
returns a filter() object which does not have a length.  To get the
length of a filter, one needs to convert it into a list which totally
defeats the purpose of filter.

This commit replace the filter used when parsing the list of sites
with a list comprehension.