Refactor objective devices, handlers, and create Objectives interface (#456)
The configuration of an objective is too complicated and in part stems from each
objective device actually being a set of objectives.  Across cockpit, most of
the GUI directly accesses the handlers via depot.

This commit rewrites the ObjectiveDevice class so that it represents a single
objective device; moves the definition of dummy objectives to cockpit.depot;
rewrites the ObjectiveHandler to correspond to a single objective; creates a
new Objectives interface that handle changing objective and access to all the
handlers; rewrites the ObjectiveControls in mainPanels which now handles a
single Objectives interface instead of multiple ObjectiveHandlers; accommodates
the new ObjectiveControls in the mainPanel to always be in the top row since it
now never has multiple choice UI; adjusts all the code that handles objectives
to use the new Objectives interface; adds documentation for ObjectiveDevice;
adds tests.

The new Objectives interface is a wx.EvtHandler instance and places like
experiment and dataSaver now need wx to get the App Objectives singleton.  This
usage of wx should not happen, but the issue is that those classes should have
been passed the objective information they need to operate, instead of looking
for some global instance.