depot.getHandlerWithName: remove fallback to case insensitive search.
If there is no handler with the given name, it does a case insensitive search.
This can have surprising behaviour if more than one handler have the same name
(after converting to lower case).  It is also surprising that getHandlerWithName
returns a handler whose name property is not equal to the name used to get it.
And since getHandlerWithName is also used to check if a handler with such name
exists, and the case insensitive map is created new each time, that's just a
waste.  If the casing is wrong, the sooner the typo is identified, the better so
the user can fix it.

This commit removes the attempt to guess the wanted handler name by doing a case
insensitive search.