Channels: do not use cockpit events to get/apply channel settings.
The channels interface knows the type of device handlers it should be saving, so
no need to send a global event.  We plan to eventually not have a singleton
depot so sending events global will not work, we will need to get to the
individual handlers through a depot instance.

This commit removes the publishing of 'save exposure settings' and 'load
exposure settings', replacing with calling the respective methods on the
individual handlers.  It also simplifies those methods in the handlers since the
global channel dict is not passed, only the handler specific settings.  It also
enables the validation discussed on #633.