cockpit/devices: release Pyro4 proxy connection in Device.onExit (#674)
If a Pyro4 connection is not released when the python program ends, the Proxy
destructor errors (this may be a bug in cockpit itself, we have too many
singletons and cyclic references).  This issue was not visible before because we
called os._exit.

This commit explicitly releases the proxy connection to the Pyro daemon.