MicroscopeFilter: drop attempt to get_filters from remote device.
Microscope's FilterWheel no longer has get_filters method, this is
expected to be handled by the user interface software.  As it was,
even filter definition was missing from cockpit configuration, the
error message would be about a missing `get_filters` method when it
should really be about the missing config.