BoulderSLM.examineActions: fix check for actions already processed (#647)
Actions already processed, at least for the boulder SLM device, will
have a True or False value.  The condition being removed was meant to
ignore actions that have already been processed.  However, check if a
value is in `[True, False]` tests for equality so it was also skipping
actions with a value of zero (it equals False).

This commit fixes by testing for identity.