MicroscopeCamera: do not use the camera.make_safe method.
A `microscope.abc.Camera` object no longer has a `make_safe` method.
Even when it did have a `make_safe` method, not all classes
implemented it.  In the ones that did implement it, it was the same as
calling `disable`.  Since we are already calling `disable` so we can
just remove the call to `make_safe`.