doc: duplicate README and INSTALL in doc and change html styling.
There are now paths for images in the README file so it looks pretty
on github.  However, the path for github is relative to the file
itself while when building the documentation with sphinx is relative
to doc root.  In addition, the icon at the top of README looks awful
on the html documentation because there's already a large icon on the
sidebar.  So now we just copy the content to doc/index to make those
tweaks.

Also, change the HTML styling to look prettier.