ExperimentConfigPanel: rewrite control to select data filepath (#401)
This commit rewrites the section of the experiment dialog about
choosing the data filepath.  It adds a control to choose the
directory, a filename template, and the actual filename.

1) It is now possible to select a directory to save the data.
Previously, this was only set via the cockpit config file so changing
this required restarting Cockpit with a different config file.

2) Filename and Template are separate so that template can have other
keys (only date and time, but others in the future) anywhere except of
only date and only at the start.

3) Because Filename and Template are separate, a separate box for the
file suffix is no longer needed.