touchscreen: display message dialog if starting mosaic with camera disabled.
If the mosaic is started from the touchscreen window with the camera
disabled, the mosaic is "cancelled" later in MosaicWindow.mosaicLoop.
While that is reflected on the state of the button itself, the user
does not get any feedback other than a message in the logging panel.

This commits has the touchscreen display a dialog message.