doc: update cockpit.devices documentation (#720)
The config files in the `sample-configs` directory was out of date.
This commit removed all of them but moved parts to the corresponding
device classes docstrings.

The documentation for the different cockpit devices corresponds to the
class and not to the module since that's what one specifices on the
config file and a module may have more than one device.  This commits
moves the documentation from the module level to the corresponding
class.

Documentation should be in the docstrings and not comments since we
are using sphinx to generate it.  This comit converts the comments
documentation into docstrings.