MosaicWindow: set wanted size of MosaicCanvas via the constructor (#692)
The MosaicCanvas does not have a method to estimate its best size.
This was being done by setting the MosaicWindow client size and the
MosaicCanvas should then expand to fill its space.  However, since
wxPython 4.1.1 in Linux, this fails.  While SetClientSize works, when
the window is show it is shrunk back.  Not sure why.  Setting an
initial size when constructing the canvas seems to fix it.