cockpit.gui.freetype: delete glyph textures when window is destroyed (#684)
The GL textures were being deleted via the Python class finaliser.
This leads to a situation where they are deleted after the GL context
has already been destroyed.  This is because the GLContext is being
destroyed in the wx internals while the Glyph finaliser is being
called after the wx.window object is destroyed by Python.

This commit adds a release method to Glyph and Face so that they can
be called manually.  The Face class now takes the window argument so
that it can bind the release to the window destruction (the
alternative is that each window does it itself which is a lot more
verbose).