maint: assign copyright to institutions and not individual authors (#722)
The work done by University of Oxford employees should be assigned to
the university.  This commit does that, i.e. assigns work done by
Mick, David, Ian, and Thomas to the University of Oxford.

Work done by PhD students belongs to the student so Nicholas Hall and
Frederik Hall keep it.  Danail Stoychev did not want copyright
assigned to him and requested it to be assigned to the University of
Oxford.

This commit also adds an AUTHORS file to describe contributions of
different contributors.