Require Pyro4 version 4.22 or later.
We use the distutils.version package to check the Pyro4 version and
change some settings if using Pyro4 older than 4.22.  However,
distutils.version is undocumented, distutils has been deprecated for
many years, and Python 3.12 will remove distutils.  In addition, Pyro
4.22 was released in August 2023, almost 10 years ago.

Instead of changing the way the way we check the version number, this
commit adds a requirement for Pyro 4.22 or later.