MicroscopeStage.enable: minor style changes.
Move block defining variables near to where they are use (which moves
the block enabling the stage further up.

Remove unfinished comment "this runs enable on axis that may".

Remove the condition `elif (not self._proxy.get_is_enabled())` which
just an else.  If it's already enabled, the Python-Microscope should
handle this properly and do nothing.