cockpit/depot.py: convert all functions into passthrough to class methods.
The cockpit.depot module has been used as if it's a singleton object
with its functions acting as methods of that singleton.  We do this in
many other places but there's no reason for this (maybe there was a
reason 20 years ago) and we've been slowing been moving away from it.

This commit creates new methods to the DeviceDepot for the
cockpit.depot functions which this commit turns into simple
passthroughs to the singleton object.

The plan is to next replace the singleton with property to the wx.App
and use it instead of the module function which can then be removed.