cockpit/gui/joystick.py: check wxUSE_JOYSTICK instead of skipping Macs (#870)
Support for joystick in Cockpit was disabled in Macs because it was
segfaulting in Ian's laptop (b0ac81db0de4).  However, it does work in
other Macs.  I hypothesise that the reason it failed on Ian's laptop
is that wx was built without Joystick support on it and we can check
for that ourselves.  This prevents it from failing in other platforms
which built wx without joystick and enables it to work in Macs that do
have it working.