getUserSaveDir: use config 'data-dir' without appending username.
This was making the default save dir to 'data-dir'/$username and
'data-dir' default is already on the user's own home directory.
Appending the username is redundant on the default case.  This only
makes sense when 'data-dir' is not on the home directory which means
the user is already setting 'data-dir' themselves at which point they
can rely on envs expansion to get that behaviour.

This commits removes that appending and documents that the path
expands tilde and environment variables so people can easily go back
to the previous behaviour.