getUserSaveDir: replace with reading straight from App.Config.
getUserSaveDir function is from back before we had a central
configuration on the wxApp itself.  It now simply reads it at start.
We might as well read it from Config directly which starts support to
change 'data-dir' during a session.

This commit removes getUserSaveDir and the whole cockpit.util.files
which now has nothing left.