Configure logging as early as possible as configurations are read.
We currently only configure logging after parsing command line args
and configuration files.  So we have no logs for that process.  This
commit changes it so that we configure logging as we read the
configuration.

This commit does the minimal standard logging configuration right at
the start, and moves the parsing of command lines outside of
CockpitConfig so that we can apply --debug early.