cockpit/util/Mrc.py: set DVID with decimal numbers instead of 0xc0a0 (#896)
The MRC/DV file header has a DVID field which should be set to -16224.
This has always been set with 0xc0a0 (49312) and this being a 16bit
signed int then wraps around to -16224.  Numpy 2 no longer wraps
around signed integers ("conversion of out-of-bound Python integers"
was deprecated in Numpy 1.24 and is an error since Numpy 2.0).

This commit simply replaces the hexadecimal value to the expected
value in decimal format.