cockpit.gui.device.Menu: remove barely used class (#910)
This Menu class is an util meant to ease creation of "dialogs" that
are a series of buttons.  In practice, it is only used by BoulderSLM
and using it requires an extra layer that makes things more convoluted
(see BoulderSLM.menuCallback).

The implementation has the issue of using ID zero to menu items which
does not work on MacOS (see issue #910).  Instead of fixing the use of
IDs zero, this commit simply removes the Menu class and BoulderSLM now
creates its own Menu directly in a simpler manner.